requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.31
alembic>=1.13.2
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.0
python-multipart>=0.0.9
pydantic-settings>=2.3.4

# Testing
pytest>=8.2.2
httpx>=0.27.0
